Move ConditionBase into dedicated header file
