+++ parse/FieldsParser.cpp
@@ -3,7 +3,6 @@
 #include "ParseImpl.h"
 #include "EffectParser.h"
 
-#include "../universe/Condition.h"
 #include "../universe/Effect.h"
 #include "../universe/Field.h"
