+++ parse/ShipHullsParser.cpp
@@ -10,7 +10,6 @@
 #include "CommonParamsParser.h"
 
 #include "../universe/ValueRef.h"
-#include "../universe/Condition.h"
 #include "../universe/ShipDesign.h"
 
 #include <boost/spirit/include/phoenix.hpp>
