+++ python/server/ServerFramework.cpp
@@ -5,6 +5,7 @@
 #include "../CommonWrappers.h"
 
 #include "../../util/Logger.h"
+#include "../../universe/ConditionBase.h"
 #include "../../universe/Universe.h"
 #include "../../universe/UniverseGenerator.h"
