+++ universe/Fighter.h
@@ -2,7 +2,7 @@
 #define _Fighter_h_
 
 #include "UniverseObject.h"
-#include "../universe/Condition.h"
+#include "../universe/ConditionBase.h"
 #include "../util/Export.h"
 
 ////////////////////////////////////////////////
