+++ universe/ValueRef.h
@@ -2,7 +2,7 @@
 #define _ValueRef_h_
 
 #include "ScriptingContext.h"
-#include "Condition.h"
+#include "ConditionBase.h"
 #include "Universe.h"
 #include "../util/Export.h"
 #include "../util/i18n.h"
@@ -12,6 +12,7 @@
 #include <boost/algorithm/string/case_conv.hpp>
 #include <boost/lexical_cast.hpp>
 #include <boost/format.hpp>
+#include <boost/serialization/nvp.hpp>
 
 #include <iterator>
 #include <map>
